Added findsum() and contains() methods
findSum() adds all the values together using a 'runner' to iterate through the list in our nodes and returns the sum. contains() will also use the 'runner' to go through the list and check if the value provided as an argument to our contains() method exists in our list. Returns true or false.